- moved most of remoterss/RemoteRssActivity code into RssfeedsActivity - RemoteRssFragment now extends android.support.v4.app.Fragment - removed swipeRefreshLayout
