+++ app/src/main/res/layout/fragment_remoterss.xml
@@ -27,16 +27,11 @@
             android:gravity="center"
             android:id="@+id/remoterss_status_message"/>
 
-        <android.support.v4.widget.SwipeRefreshLayout
-            android:id="@+id/swipe_refresh_layout"
+        <ListView
+            android:id="@+id/torrents_list"
             android:layout_width="match_parent"
-            android:layout_height="match_parent">
-                <ListView
-                    android:id="@+id/torrents_list"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:clipToPadding="false"
-                    tools:listitem="@layout/list_item_remoterssitem"
-                    tools:visibility="visible"/>
-        </android.support.v4.widget.SwipeRefreshLayout>
+            android:layout_height="match_parent"
+            android:clipToPadding="false"
+            tools:listitem="@layout/list_item_remoterssitem"
+            tools:visibility="visible"/>
 </FrameLayout>
